refactor(engine): add assets sample dir and improve text sample
