+++ engine/samples/assets/bridge/assets/sample.txt
